requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
discord.py==1.3.4
idna==2.10
lxml~=4.5.2
multidict==4.7.6
numpy==1.19.0
pandas~=1.0.5
python-dateutil==2.8.1
pytz==2020.1
requests~=2.24.0
six==1.15.0
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
html5lib~=1.1
beautifulsoup4~=4.9.1